change emmaclient export files for relays
