+++ quickevent/app/quickevent/plugins/Relays/src/relaysplugin.h
@@ -44,6 +44,9 @@ public:
 
 	qf::core::utils::TreeTable nLegsResultsTable(const QString &where_option, int leg_count, int places, bool exclude_not_finish);
 	qf::core::utils::TreeTable nLegsClassResultsTable(int class_id, int leg_count, int places, bool exclude_not_finish);
+	QString startListIofXml30();
+	QString resultsIofXml30();
+	QVariant startListByClassesTableData(const QString &class_filter);
 private:
 	Q_SLOT void onInstalled();
 	void onDbEventNotify(const QString &domain, int connection_id, const QVariant &data);
